tool service proxy test case
